*GoogleTest added *Some basic tests *Some huge fixes :)

+++ TwitchXX/TwitchGame.h
@@ -34,18 +34,18 @@ namespace TwitchXX
 		}
 
 		std::wstring Name() const { return _name; }
-		unsigned int Channels() const { return _channels; }
-		unsigned int Viewers() const { return _viewers; }
-		unsigned int Id() const { return _id; }
-		unsigned int Giantbomb_Id() const { return _giantbomb_id; }
+		size_t Channels() const { return _channels; }
+		size_t Viewers() const { return _viewers; }
+		size_t Id() const { return _id; }
+		size_t Giantbomb_Id() const { return _giantbomb_id; }
 		const ImageCollection& Box() const { return _box; }
 		const ImageCollection& Logo() const { return _logo; }
 
 		void Name(const std::wstring& name) { _name = name; }
-		void Channels(int count) { _channels = count; }
-		void Viewers(int count) { _viewers = count; }
-		void Id(int id) { _id = id; }
-		void Giantbomb_Id(int id) { _giantbomb_id = id; }
+		void Channels(size_t count) { _channels = count; }
+		void Viewers(size_t count) { _viewers = count; }
+		void Id(size_t id) { _id = id; }
+		void Giantbomb_Id(size_t id) { _giantbomb_id = id; }
 		void Box(const ImageCollection& col) { _box = col; }
 		void Box(ImageCollection&& col) { _box = std::move(col); }
 		void Logo(const ImageCollection& col) { _logo = col; }
@@ -55,10 +55,10 @@ namespace TwitchXX
 
 	private:
 		std::wstring _name;
-		unsigned int _channels;
-		unsigned int _viewers;
-		unsigned int _id;
-		unsigned int _giantbomb_id;
+		size_t _channels;
+		size_t _viewers;
+		size_t _id;
+		size_t _giantbomb_id;
 		ImageCollection _box;
 		ImageCollection _logo;
